apps/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Python backend for the Electron React Python Template"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "pydantic-settings",
    "httpx",
]

[project.optional-dependencies]
dev = [
    "pytest",
]
